requirements.txt
Django~=5.1.1
django-modeltranslation~=0.19.9
djangorestframework~=3.15.2
django-filter~=24.3
pip~=24.2
pytz~=2024.2
PyYAML~=6.0.2
asgiref~=3.8.1
sqlparse~=0.5.1
Jinja2~=3.1.4
requests~=2.32.3
MarkupSafe~=2.1.5
certifi~=2024.8.30
itypes~=1.2.0
coreapi~=2.3.3
python-dotenv==1.0.1
coreschema~=0.0.4
uritemplate~=4.1.1
urllib3~=2.2.3
packaging~=24.1
inflection~=0.5.1
idna~=3.10
simplejson~=3.19.3
setuptools~=75.1.0